Third modification // menu, fileOperations

+++ src/App.java
@@ -12,7 +12,7 @@ public class App {
     private String path = Paths.get(".").toAbsolutePath().normalize().toString()+"/src";
     private FileWriter stream = null;
     private TreeSet<String> sortedList = new TreeSet<>();
-    private Scanner scanner = new Scanner(System.in);
+    private static Scanner scanner;
 
 
 
@@ -26,17 +26,20 @@ public class App {
 
         for(File list: this.fileList)
             sortedList.add(list.getName());
-        System.out.println("----------------------------File List----------------------------------------");
+        System.out.println("---------------------------- File List ---------------------------------------");
         for (String value: sortedList)
-            System.out.println("Sorted List: " + value);
-        System.out.println("----------------------------End Of File List---------------------------------");
+            System.out.println("> " + value);
+        System.out.println("---------------------- *** End Of File List *** ------------------------------");
+        menu();
+
     }
 
     public void createFile(){ // Creates a file in the directory
         System.out.println("Type in the file's name to be created and hit 'enter'.");
+        scanner= new Scanner(System.in);
         String name = scanner.nextLine();
         try {
-            stream = new FileWriter(this.path + "/"+name+".txt");
+            stream = new FileWriter(this.path + "/"+name);
             System.out.println("**** A file named ' " +name + " ' has been created ******\r");
             this.getFileList();
         } catch (IOException e) {
@@ -44,40 +47,58 @@ public class App {
         }finally {
             try {
                 stream.close();
+                menu();
             } catch (IOException e) {
                 e.printStackTrace();
             }
         }
     }
 
-    public void removeElement(){ // Removes a specified file
-        System.out.println("Type in the file's name to be deleted then hit 'enter'.");
+    public void removeElement() { // Removes a specified file
+        System.out.println("Type in the file's name to be deleted then hit 'enter' / type 'cancel' to return to Home.");
+        scanner = new Scanner(System.in);
         String name = scanner.nextLine();
 
-        file = new File(path +"/"+ name);
-        if(file.exists())
-            if(file.delete()) {
-                System.out.println("Deleted the file: " + file.getName());
-                getFileList();
+        if(name.equals("cancel") ) {
+            menu();
+        }else {
+            file = new File(path + "/" + name);
+            if (file.exists()) {
+                if (file.delete()) {
+                    System.out.println("Deleted the file: " + file.getName());
+                    getFileList();
+                    menu();
+                } else {
+                    System.out.println("Failed to delete the file.");
+                    menu();
+                }
             } else {
-                System.out.println("Failed to delete the file.");
+                System.out.println("File does not exist / misspelled, consider case sensitivity and file extension");
+                menu();
             }
-        else System.out.println("File does not exist / misspelled, consider case sensitivity and file extension");
-
+        }
     }
 
     public void search(){
-        System.out.println("Type a file's name then hit 'enter'.");
+        System.out.println("Type a file's name with its extension(.txt, .png, etc) then hit 'enter' / type 'cancel' to return to Home.");
+        scanner = new Scanner(System.in);
         String name = scanner.nextLine();
+        if(name.equals("cancel") ) {
+            menu();
+        }else {
         file = new File(path +"/"+ name);
 
         if(file.exists()){
             System.out.println("--------------------------------  Found File  -------------------------------");
             System.out.println(file.getName());
             System.out.println("---------------------------- **** Found File **** ---------------------------");
+            fileOperations();
         }else{
             System.out.println("--------------------------------  File Not Found  ---------------------------");
             System.out.println("No element named " + name);
+            System.out.println("---------------------------- **** File Not Found **** -----------------------");
+            search();
+        }
         }
     }
 
@@ -91,18 +112,25 @@ public class App {
         System.out.println(" --------------------------------------------------------------------------------------");
         System.out.println("");
         System.out.println("");
-        menu();
+        try {
+            menu();
+        }catch ( Exception e){
+            System.out.println("Bad input try again");
+            menu();
+        }
+
 
     }
 
 
     public void menu(){
 
-        System.out.println("1) List all files in the current directory.");
-        System.out.println("2) File Operations (Add , Delete and Search.)   ");
-        System.out.println("3) Home.  ");
-        System.out.println("4) Exit.  ");
+        System.out.println("1) List all files in the current directory");
+        System.out.println("2) File Operations (Add , Delete and Search)   ");
+        System.out.println("3) Home  ");
+        System.out.println("4) Exit  ");
 
+        scanner = new Scanner(System.in);
         int option = scanner.nextInt();
 
         switch (option){
@@ -110,9 +138,9 @@ public class App {
             case 1 :
                 getFileList();
                 break;
-            case 2 :{
-                fileOperations();
-                break;}
+            case 2 :
+                    fileOperations();
+                break;
             case 3:
                 load();
             case 4 :
@@ -123,26 +151,42 @@ public class App {
     }
 
     public void fileOperations(){
-        System.out.println("1) Add file to the current Directory");
-        System.out.println("2) Delete a file ");
-        System.out.println("3) Search a file ");
+        try {
+            System.out.println("1) Add file to the current Directory");
+            System.out.println("2) Delete a file ");
+            System.out.println("3) Search a file ");
+            System.out.println("4) Home  ");
+            System.out.println("5) Exit  ");
 
-        int options = 0;
-        options = scanner.nextInt();
+            scanner = new Scanner(System.in);
+            int options = scanner.nextInt();
 
-        switch (options){
-            case 1 :
-                createFile();
-                break;
-            case 2 :
-                removeElement();
-                break;
-            case 3 :
-                search();
-            default:
-                System.out.println("Option does not exist");
+            switch (options){
+                case 1 :
+                    createFile();
+                    break;
+                case 2 :
+                    removeElement();
+                    break;
+                case 3 :
+                    search();
+                    break;
+                case 4 :
+                    load();
+                    break;
+                case 5 :
+                    System.exit(0);
+                    break;
+                default:
+                    System.out.println("Option does not exist");
+                    fileOperations();
 
+            }
+        }catch (Exception e){
+            System.out.println("Bad input try again..");
+            fileOperations();
         }
+
 
     }
 }
